+++ graphics/tuple.go
@@ -2,8 +2,8 @@ package graphics
 
 //Tuple struct
 type Tuple struct {
-	X float64
-	Y float64
-	Z float64
+	X float32
+	Y float32
+	Z float32
 	W int
 }
